Move code around to make things easier to follow
