+++ .gitignore
@@ -128,6 +128,8 @@ dmypy.json
 # Pyre type checker
 .pyre/
 
+.idea/
+
 
 config.py
 *.sqlite
